requirement.txt
celery==5.2.7
Django==4.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-ckeditor-5==0.2.2
django-cors-headers==3.13.0
django-cryptography==1.1
django-debug-toolbar==4.0.0
django-debug-toolbar-force==0.2
django-elasticsearch-debug-toolbar==3.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-filter==22.1
django-ibanfield==0.0.2
django-lifecycle==1.0.0
django-nine==0.2.7
django-redis==5.2.0
django-rest-swagger==2.2.0
django-storages==1.13.1
django-timezone-field==5.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
psycopg2==2.9.9
gunicorn==21.2.0
boto3==1.28.78
matplotlib==3.8.1
google-api-core==2.14.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
python-dotenv==1.0.0